Adjust split direction based on how wide or tall the current shape is
Ie if a shape is twice as tall as it is wide, multiply the likelyhood of all the
row centric splits by 1.5, and all the column centric splits by 0.66. That way
it's still possible to split repeatedly by row or column but it gets
increasingly less likely to split too far in one direction.