Add floating point fix to handle extended lines
I *think* what was happening is that the solution for xmax,y(xmax) and
x(ymax),ymax were close enough to the boundary it would oscillate triggering
high-x or high-y and never converge. Ensuring the upper bound is always within
the bounds by a tolerance threshold appears to resolve this, as smaller
tolerances trigger this again.