Fix non-determinism in dr/weighted using sequential ordering
dr/weighted had a bug if the weights are in a map. Changes dr/weighted to *both*
support a map or a sequential pairing of values to weights.

As radial-symmetries was using a set of anonymous functions, sometimes the sort
order of the map would change, forcing non-determinism into the sketch.