handle an edge case that shouldn't be possible
Not sure cause here, probably some kind of incorrect generation of the vertices
in a polygon, but was throwing errors about nil as an input to gv/vec2 otherwise
and showing up as a spooky action at a distance. c1 is definitely not right but
it's vastly preferable than the current error situation.